# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CityConnect API
  version: "1.0.0"
  description: |
    Municipal CRM and service-request API.

    Three kinds of caller reach this API, and they authenticate differently:

    - **Staff** hold an HTTP-only session cookie, established by signing in
      through C2 (`/auth/login` → C2 → `/auth/callback`). There is no local
      account path.
    - **Connected systems** present a `cc_pat_…` personal access token as a
      bearer credential. A token can only ever narrow its owner's authority.
    - **C2 itself** calls two endpoints server-to-server: the Service Card
      callout (`/citizens/{sub}/status`, authenticated by a short-lived RS256
      assertion) and back-channel logout (a signed `logout_token`).

    Errors are `application/problem+json` and carry a stable `code` field.
    Switch on that rather than on the prose.
  contact:
    name: CityConnect
    url: https://github.com/jjamieson1/CityConnect
servers:
  - url: https://city.example.gov/cityconnect/api
    description: Production
  - url: http://localhost:4021/api
    description: Local development
...
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: hasC2Link, in: query, schema: { type: boolean } }
        - { name: c2Reachable, in: query, schema: { type: boolean } }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Contact" } }
    post:
      tags: [Contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Contact" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
  /contacts/{id}/merge:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [Contacts]
      summary: Merge a duplicate into this contact
      description: |
        Moves requests, interactions, identities and preferences onto the
        survivor and retains the loser so its id keeps resolving. Reversible
        from the merge history — an irreversible merge is how a CRM loses a
        citizen's history permanently.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mergedId]
              properties:
                mergedId: { type: string }
                note: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }

  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cc_session
      description: Set by the C2 sign-in callback. HTTP-only, SameSite=Lax.
    bearerToken:
      type: http
      scheme: bearer
      description: A `cc_pat_…` personal access token, for connected systems.
    calloutAssertion:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        C2's short-lived RS256 Service Card assertion, verified against the same
        JWKS as id_tokens. `aud` must equal our `client_id`; `exp` is about 60
        seconds after `iat`.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
...
  schemas:
...
    Page:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        hasMore: { type: boolean }
...
    Contact:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        primaryEmail: { type: string }
        primaryPhone: { type: string }
        doNotContact: { type: boolean, description: "A hard stop on all outbound messaging." }
        c2Reachable:
          type: boolean
          description: |
            Maintained from C2's responses. False means a notification was
            refused — consent withdrawn or the identity is unknown — so this
            citizen must be reached by phone or post.
        version: { type: integer }
...